Added more float accuracies
Changed rockets and grenades into floats for accuracy
Fixed math pertaining to playerAccuracy again

+++ src/main/java/stuuupiiid/guncus/entity/EntityGrenade.java
@@ -43,7 +43,7 @@ public class EntityGrenade extends EntityProjectile implements IProjectile, IEnt
 		canBePickedUp = 0;
 	}
 	
-	public EntityGrenade(World parWorld, EntityPlayer entityPlayer, double speed, int accuracy) {
+	public EntityGrenade(World parWorld, EntityPlayer entityPlayer, double speed, float accuracy) {
 		super(parWorld, entityPlayer, speed, accuracy);
 		
 		setSize(0.5F, 0.5F);
